feat: implement FAQ registration page with form validation and image upload
