+++ package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.0.1",
+    "@hookform/resolvers": "^5.1.1",
     "@tailwindcss/vite": "^4.1.4",
     "ci": "^2.3.0",
     "clsx": "^2.1.1",
@@ -25,7 +25,7 @@
     "swiper": "^11.2.8",
     "tailwind-merge": "^3.2.0",
     "vite-tsconfig-paths": "^5.1.4",
-    "zod": "^3.25.28"
+    "zod": "^3.25.67"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
